pyproject.toml
[project]
name = "teste"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "fastapi>=0.135.1",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "pytz>=2026.1.post1",
    "rich>=14.3.3",
    "uvicorn>=0.41.0",
]
